refactor(sandbox-fc): move control server tests
